requirements.txt
BeautifulSoup4
requests
pandas
datetime
numpy
mibian
dash
plotly.express